backend/pyproject.toml
[project]
name = "supoclip-backend"
version = "0.1.0"
description = "Python-based backend for SupoClip"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "openai-whisper>=20250625",
    "pydantic-ai>=0.4.9",
    "setuptools-rust>=1.11.1",
    "aiosqlite>=0.20.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2025.7.21",
    "whisper>=1.1.10",
    "greenlet>=2.0.0",
    "moviepy>=2.2.1",
    "opencv-python>=4.8.0",
    "srt-equalizer>=0.1.10",
    "assemblyai>=0.56.0",
    "srt>=3.5.3",
    "mediapipe>=0.10.0",
    "numpy>=1.24.0",
    "aiofiles>=23.2.0",
    "sse-starlette>=1.6.5",
    "arq>=0.26.0",
    "redis>=5.0.0",
    "groq>=0.9.0",
    "edge-tts>=7.2.7",
    "duckdb>=0.9.2",
    "httpx>=0.27.0",
    "google-generativeai>=0.4.0",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
